PF-57 add: add component MyPayments, Table and types
