+++ src/lib/feature/profile/api/profileApi.ts
@@ -37,9 +37,6 @@ export const profileApi = baseApi.injectEndpoints({
       },
       keepUnusedDataFor: 0,
     }),
-    getMyPayments: build.query<GetMyPayments[], void>({
-      query: () => '/subscriptions/my-payments',
-    }),
   }),
 })
 
@@ -48,15 +45,4 @@ export const {
   useGetUserPostsQuery,
   useGetProfileQuery,
   useUpdateProfileMutation,
-  useGetMyPaymentsQuery,
 } = profileApi
-
-export type GetMyPayments = {
-  dateOfPayment: string
-  endDateOfSubscription: string
-  paymentType: 'STRIPE' | 'PAYPAL' | 'CREDIT_CARD'
-  price: number
-  subscriptionId: string
-  subscriptionType: 'DAY' | 'WEEKLY' | 'MONTHLY'
-  userId: number
-}
