+++ src/lib/feature/profile/ui/components/profile/MyPayments/MyPayments.tsx
@@ -1,8 +1,8 @@
 'use client'
 
 import { Table } from '@/lib/feature/profile/ui/components/profile/MyPayments/Table/Table'
-import { useGetMyPaymentsQuery } from '@/lib/feature/profile/api/profileApi'
 import Loader from '@/components/ui/loader/Loader'
+import { useGetMyPaymentsQuery } from '@/lib/feature/subscriptions/api/subscriptionApi'
 
 export const MyPayments = () => {
   const { data, isLoading } = useGetMyPaymentsQuery()
