+++ src/lib/feature/subscriptions/api/subscriptionApi.types.ts
@@ -17,3 +17,13 @@ export type getCurrentSubscriptionResponse = {
   data: Data[]
   hasAutoRenewal: boolean
 }
+
+export type GetMyPayments = {
+  dateOfPayment: string
+  endDateOfSubscription: string
+  paymentType: 'STRIPE' | 'PAYPAL' | 'CREDIT_CARD'
+  price: number
+  subscriptionId: string
+  subscriptionType: 'DAY' | 'WEEKLY' | 'MONTHLY'
+  userId: number
+}
